Kernel: Use u64 instead of size_t for File::can_read offset
This ensures offsets will not be truncated on large files on i686.

+++ Kernel/Bus/VirtIO/ConsolePort.cpp
@@ -85,7 +85,7 @@ void ConsolePort::handle_queue_update(Badge<VirtIO::Console>, u16 queue_index)
     }
 }
 
-bool ConsolePort::can_read(const OpenFileDescription&, size_t) const
+bool ConsolePort::can_read(const OpenFileDescription&, u64) const
 {
     return m_receive_buffer->used_bytes() > 0;
 }
